requirements.txt
experta==1.9.4
fastapi==0.79.0
pydantic>=1.8.0
uvicorn>=0.15.0
jinja2==3.1.2
python-multipart==0.0.5
python-dotenv==0.21.0